Allow custom video title to be set.
On iOS devices, if NoSleep.js is running and you close the display and
reopen it, you will see the name of the video running. For some users,
this name can be confusing as to them, "No Sleep" is not intuitive. We
provide the ability now to customize this title to match your app.